improved web refreshing performance
